langchain-agent/requirements.txt
xenarch[agent,langchain]
langchain>=0.3.0
langchain-openai>=0.3.0
python-dotenv
